Move value to new file
