+++ src/event/mod.rs
@@ -5,22 +5,23 @@ use getset::{Getters, Setters};
 use lazy_static::lazy_static;
 use metric::{MetricKind, MetricValue};
 use once_cell::sync::OnceCell;
-use serde::{Deserialize, Serialize, Serializer};
-use serde_json::Value as JsonValue;
+use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 use string_cache::DefaultAtom as Atom;
 
 pub mod discriminant;
+mod log_event;
 pub mod merge;
 pub mod merge_state;
 pub mod metric;
 mod util;
-mod log_event;
+mod value;
 
+pub use log_event::LogEvent;
 pub use metric::{Metric, StatisticKind};
 pub(crate) use util::log::PathComponent;
 pub(crate) use util::log::PathIter;
-pub use log_event::LogEvent;
+pub use value::Value;
 
 pub mod proto {
     include!(concat!(env!("OUT_DIR"), "/event.proto.rs"));
@@ -138,192 +139,6 @@ impl LogSchema {
     }
     fn default_source_type_key() -> Atom {
         Atom::from("source_type")
-    }
-}
-
-#[derive(PartialEq, Debug, Clone)]
-pub enum Value {
-    Bytes(Bytes),
-    Integer(i64),
-    Float(f64),
-    Boolean(bool),
-    Timestamp(DateTime<Utc>),
-    Map(BTreeMap<String, Value>),
-    Array(Vec<Value>),
-    Null,
-}
-
-impl Serialize for Value {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        match &self {
-            Value::Integer(i) => serializer.serialize_i64(*i),
-            Value::Float(f) => serializer.serialize_f64(*f),
-            Value::Boolean(b) => serializer.serialize_bool(*b),
-            Value::Bytes(_) | Value::Timestamp(_) => {
-                serializer.serialize_str(&self.to_string_lossy())
-            }
-            Value::Map(m) => serializer.collect_map(m),
-            Value::Array(a) => serializer.collect_seq(a),
-            Value::Null => serializer.serialize_none(),
-        }
-    }
-}
-
-impl From<Bytes> for Value {
-    fn from(bytes: Bytes) -> Self {
-        Value::Bytes(bytes)
-    }
-}
-
-impl From<Vec<u8>> for Value {
-    fn from(bytes: Vec<u8>) -> Self {
-        Value::Bytes(bytes.into())
-    }
-}
-
-impl From<&[u8]> for Value {
-    fn from(bytes: &[u8]) -> Self {
-        Value::Bytes(Vec::from(bytes).into())
-    }
-}
-
-impl From<String> for Value {
-    fn from(string: String) -> Self {
-        Value::Bytes(string.into())
-    }
-}
-
-impl From<&String> for Value {
-    fn from(string: &String) -> Self {
-        string.as_str().into()
-    }
-}
-
-impl From<&str> for Value {
-    fn from(s: &str) -> Self {
-        Value::Bytes(Vec::from(s.as_bytes()).into())
-    }
-}
-
-impl From<DateTime<Utc>> for Value {
-    fn from(timestamp: DateTime<Utc>) -> Self {
-        Value::Timestamp(timestamp)
-    }
-}
-
-impl From<f32> for Value {
-    fn from(value: f32) -> Self {
-        Value::Float(f64::from(value))
-    }
-}
-
-impl From<f64> for Value {
-    fn from(value: f64) -> Self {
-        Value::Float(value)
-    }
-}
-
-impl From<BTreeMap<String, Value>> for Value {
-    fn from(value: BTreeMap<String, Value>) -> Self {
-        Value::Map(value)
-    }
-}
-
-impl From<Vec<Value>> for Value {
-    fn from(value: Vec<Value>) -> Self {
-        Value::Array(value)
-    }
-}
-
-macro_rules! impl_valuekind_from_integer {
-    ($t:ty) => {
-        impl From<$t> for Value {
-            fn from(value: $t) -> Self {
-                Value::Integer(value as i64)
-            }
-        }
-    };
-}
-
-impl_valuekind_from_integer!(i64);
-impl_valuekind_from_integer!(i32);
-impl_valuekind_from_integer!(i16);
-impl_valuekind_from_integer!(i8);
-impl_valuekind_from_integer!(isize);
-
-impl From<bool> for Value {
-    fn from(value: bool) -> Self {
-        Value::Boolean(value)
-    }
-}
-
-impl From<JsonValue> for Value {
-    fn from(json_value: JsonValue) -> Self {
-        match json_value {
-            JsonValue::Bool(b) => Value::Boolean(b),
-            JsonValue::Number(n) => {
-                if let Some(i) = n.as_i64() {
-                    Value::Integer(i)
-                } else if let Some(f) = n.as_f64() {
-                    Value::Float(f)
-                } else {
-                    Value::Bytes(n.to_string().into())
-                }
-            }
-            JsonValue::String(s) => Value::Bytes(Bytes::from(s)),
-            JsonValue::Object(obj) => Value::Map(
-                obj.into_iter()
-                    .map(|(key, value)| (key, Value::from(value)))
-                    .collect(),
-            ),
-            JsonValue::Array(arr) => Value::Array(arr.into_iter().map(Value::from).collect()),
-            JsonValue::Null => Value::Null,
-        }
-    }
-}
-
-impl Value {
-    // TODO: return Cow
-    pub fn to_string_lossy(&self) -> String {
-        match self {
-            Value::Bytes(bytes) => String::from_utf8_lossy(&bytes).into_owned(),
-            Value::Timestamp(timestamp) => timestamp_to_string(timestamp),
-            Value::Integer(num) => format!("{}", num),
-            Value::Float(num) => format!("{}", num),
-            Value::Boolean(b) => format!("{}", b),
-            Value::Map(map) => serde_json::to_string(map).expect("Cannot serialize map"),
-            Value::Array(arr) => serde_json::to_string(arr).expect("Cannot serialize array"),
-            Value::Null => "<null>".to_string(),
-        }
-    }
-
-    pub fn as_bytes(&self) -> Bytes {
-        match self {
-            Value::Bytes(bytes) => bytes.clone(), // cloning a Bytes is cheap
-            Value::Timestamp(timestamp) => Bytes::from(timestamp_to_string(timestamp)),
-            Value::Integer(num) => Bytes::from(format!("{}", num)),
-            Value::Float(num) => Bytes::from(format!("{}", num)),
-            Value::Boolean(b) => Bytes::from(format!("{}", b)),
-            Value::Map(map) => Bytes::from(serde_json::to_vec(map).expect("Cannot serialize map")),
-            Value::Array(arr) => {
-                Bytes::from(serde_json::to_vec(arr).expect("Cannot serialize array"))
-            }
-            Value::Null => Bytes::from("<null>"),
-        }
-    }
-
-    pub fn into_bytes(self) -> Bytes {
-        self.as_bytes()
-    }
-
-    pub fn as_timestamp(&self) -> Option<&DateTime<Utc>> {
-        match &self {
-            Value::Timestamp(ts) => Some(ts),
-            _ => None,
-        }
     }
 }
 
@@ -487,7 +302,7 @@ impl From<Event> for proto::EventWrapper {
     fn from(event: Event) -> Self {
         match event {
             Event::Log(log_event) => {
-                let fields: BTreeMap::<_, _> = log_event.into();
+                let fields: BTreeMap<_, _> = log_event.into();
                 let fields = fields
                     .into_iter()
                     .map(|(k, v)| (k, encode_value(v)))
